Use the correct word list.
I cross checked this with a random sampling of cards from the
game and got a 100% match (and it's the right number of words),
so I believe this is the proper list.

Also, formatted as one word per line because my vim was really
struggling to handle all the words on a single line for some reason.