Change how configuration is stored
After a bit of thinking how best to deploy applications with secrets,
the closest thing to 12-factor methodology is to specify the production
configuration on, you guessed it, the production server. So now the
flags are specified in the Systemd service file instead of in a config
file that gets rolled into the image.