Building out AI server, and some refactoring
Because we're modelling AI clients as just normal users, we need to
update our API so clients can volunteer if they're humans or robots. Not
sure how I initially planned on doing it, but this way seems to make the
most sense.

Also, add more of the game-joining logic to the AI server.