Turns out I was using an old node:alpine image
A lot can change in four months apparently, like no longer having a
package called 'python' and forcing your users to actually specify which
version they want. Honestly it makes a lot of sense and I was surprised
it had worked (locally, not on CI) in the first place.

The other issue was a mismatched version of node-sass and Node. This
comment [1] explained it. node-sass 5 doesn't support Node 16, which is
what we were using. I've followed up with Next [2], we'll see what
happens.

In the mean time, it's now broken locally and remotely, but hopefully
updating to Sass 6 (in the future) will fix it.

[1] https://github.com/sass/node-sass/issues/2165#issuecomment-383779947
[2] https://github.com/vercel/next.js/discussions/25844